+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 * New interactive command `cider-undef`.
 * New interactive command `cider-clear-compilation-highlights`.
 * First pass at a CIDER quick reference card.
-
 * `completion-at-point` now annotates functions, macros and special forms, thus making it
 simpler to gain understanding of what you're using.
+* When invoked with a prefix argument `cider-quit` doesn't ask for confirmation.
 
 ### Changes